Refactoring: moved base/estimators to deeplearning/base/estimators.
