+++ sktime_dl/deeplearning/cnn/_classifier.py
@@ -3,7 +3,7 @@ __author__ = "James Large"
 import keras
 import numpy as np
 
-from sktime_dl.base.estimators import BaseDeepClassifier
+from sktime_dl.deeplearning.base.estimators import BaseDeepClassifier
 from sktime_dl.deeplearning.cnn._base import CNNNetwork
