+++ sktime_dl/deeplearning/inceptiontime/_classifier.py
@@ -1,7 +1,7 @@
 import keras
 import numpy as np
 
-from sktime_dl.base.estimators._classifier import BaseDeepClassifier
+from sktime_dl.deeplearning.base.estimators._classifier import BaseDeepClassifier
 
 
 class InceptionTimeClassifier(BaseDeepClassifier):
