+++ sktime_dl/deeplearning/mcnn/_classifier.py
@@ -10,7 +10,7 @@ import gc
 
 from sklearn.model_selection import train_test_split
 
-from sktime_dl.base.estimators._classifier import BaseDeepClassifier
+from sktime_dl.deeplearning.base.estimators._classifier import BaseDeepClassifier
 
 
 class MCNNClassifier(BaseDeepClassifier):
