+++ sktime_dl/deeplearning/resnet/_classifier.py
@@ -3,7 +3,7 @@ __author__ = "James Large"
 import keras
 import numpy as np
 
-from sktime_dl.base.estimators._classifier import BaseDeepClassifier
+from sktime_dl.deeplearning.base.estimators._classifier import BaseDeepClassifier
 
 
 class ResNetClassifier(BaseDeepClassifier):
