+++ sktime_dl/deeplearning/tlenet/_classifier.py
@@ -3,7 +3,7 @@ __author__ = "Aaron Bostrom, James Large"
 import keras
 import numpy as np
 
-from sktime_dl.base.estimators._classifier import BaseDeepClassifier
+from sktime_dl.deeplearning.base.estimators._classifier import BaseDeepClassifier
 
 
 class TLENETClassifier(BaseDeepClassifier):
